refactor(bot): split cloud agent session group state
