+++ apps/web/src/app/api/internal/bot-session-callback/[botRequestId]/route.ts
@@ -17,6 +17,8 @@ import {
   claimBotRequestCloudAgentSessionGroupContinuation,
   getBotRequestCloudAgentSession,
   getBotRequestCloudAgentSessionGroupReadiness,
+} from '@/lib/bot/cloud-agent-session-groups';
+import {
   markBotRequestCloudAgentSessionTerminal,
   recordBotRequestCloudAgentSessionResult,
   recordBotRequestCloudAgentSessionResultError,
