+++ apps/web/src/lib/bot/request-logging.test.ts
@@ -1,8 +1,5 @@
 import { db } from '@/lib/drizzle';
 import {
-  claimBotRequestCloudAgentSessionGroupContinuation,
-  getBotRequestCloudAgentSessionGroup,
-  getBotRequestCloudAgentSessionGroupReadiness,
   linkBotRequestToSession,
   markBotRequestCloudAgentSessionTerminal,
   recordBotRequestCloudAgentSession,
@@ -295,176 +292,6 @@ describe('bot request logging', () => {
     expect(row.final_message).toBeNull();
     expect(row.final_message_fetched_at).toBeNull();
     expect(row.final_message_error).toHaveLength(4000);
-  });
-
-  it('waits to claim a child session group until all siblings are terminal with stored results', async () => {
-    const botRequestId = await createBotRequest();
-    const spawnGroupId = randomUUID();
-    const firstCloudAgentSessionId = `cas-child-group-first-${randomUUID()}`;
-    const secondCloudAgentSessionId = `cas-child-group-second-${randomUUID()}`;
-    createdCloudAgentSessionIds.add(firstCloudAgentSessionId);
-    createdCloudAgentSessionIds.add(secondCloudAgentSessionId);
-
-    await recordBotRequestCloudAgentSession({
-      botRequestId,
-      spawnGroupId,
-      cloudAgentSessionId: firstCloudAgentSessionId,
-      callbackStep: 2,
-    });
-    await recordBotRequestCloudAgentSession({
-      botRequestId,
-      spawnGroupId,
-      cloudAgentSessionId: secondCloudAgentSessionId,
-      callbackStep: 2,
-    });
-
-    await markBotRequestCloudAgentSessionTerminal({
-      botRequestId,
-      cloudAgentSessionId: firstCloudAgentSessionId,
-      status: 'completed',
-      kiloSessionId: `kilo-first-${randomUUID()}`,
-    });
-
-    const partiallyCompleteGroup = await getBotRequestCloudAgentSessionGroup({
-      botRequestId,
-      cloudAgentSessionId: firstCloudAgentSessionId,
-    });
-    expect(partiallyCompleteGroup.sessions.map(session => session.status).sort()).toEqual([
-      'completed',
-      'running',
-    ]);
-    await expect(
-      claimBotRequestCloudAgentSessionGroupContinuation({
-        botRequestId,
-        cloudAgentSessionId: firstCloudAgentSessionId,
-      })
-    ).resolves.toBe(false);
-
-    await markBotRequestCloudAgentSessionTerminal({
-      botRequestId,
-      cloudAgentSessionId: secondCloudAgentSessionId,
-      status: 'completed',
-      kiloSessionId: `kilo-second-${randomUUID()}`,
-    });
-
-    await expect(
-      getBotRequestCloudAgentSessionGroupReadiness({
-        botRequestId,
-        cloudAgentSessionId: secondCloudAgentSessionId,
-      })
-    ).resolves.toMatchObject({ status: 'waiting-for-result' });
-    await expect(
-      claimBotRequestCloudAgentSessionGroupContinuation({
-        botRequestId,
-        cloudAgentSessionId: secondCloudAgentSessionId,
-      })
-    ).resolves.toBe(false);
-
-    await recordBotRequestCloudAgentSessionResult({
-      botRequestId,
-      cloudAgentSessionId: firstCloudAgentSessionId,
-      finalMessage: 'first final message',
-    });
-    await recordBotRequestCloudAgentSessionResult({
-      botRequestId,
-      cloudAgentSessionId: secondCloudAgentSessionId,
-      finalMessage: 'second final message',
-    });
-
-    await expect(
-      getBotRequestCloudAgentSessionGroupReadiness({
-        botRequestId,
-        cloudAgentSessionId: secondCloudAgentSessionId,
-      })
-    ).resolves.toMatchObject({ status: 'ready' });
-    await expect(
-      claimBotRequestCloudAgentSessionGroupContinuation({
-        botRequestId,
-        cloudAgentSessionId: secondCloudAgentSessionId,
-      })
-    ).resolves.toBe(true);
-    await expect(
-      claimBotRequestCloudAgentSessionGroupContinuation({
-        botRequestId,
-        cloudAgentSessionId: firstCloudAgentSessionId,
-      })
-    ).resolves.toBe(false);
-
-    const claimedGroup = await getBotRequestCloudAgentSessionGroup({
-      botRequestId,
-      cloudAgentSessionId: firstCloudAgentSessionId,
-    });
-    expect(claimedGroup.sessions).toHaveLength(2);
-    expect(claimedGroup.sessions.every(session => session.continuation_started_at)).toBe(true);
-  });
-
-  it('reports result errors and still claims the terminal group once', async () => {
-    const botRequestId = await createBotRequest();
-    const spawnGroupId = randomUUID();
-    const cloudAgentSessionId = `cas-child-group-result-error-${randomUUID()}`;
-    createdCloudAgentSessionIds.add(cloudAgentSessionId);
-
-    await recordBotRequestCloudAgentSession({
-      botRequestId,
-      spawnGroupId,
-      cloudAgentSessionId,
-    });
-    await markBotRequestCloudAgentSessionTerminal({
-      botRequestId,
-      cloudAgentSessionId,
-      status: 'completed',
-      kiloSessionId: `kilo-result-error-${randomUUID()}`,
-    });
-    await recordBotRequestCloudAgentSessionResultError({
-      botRequestId,
-      cloudAgentSessionId,
-      errorMessage: 'ingest returned no final assistant text',
-    });
-
-    await expect(
-      getBotRequestCloudAgentSessionGroupReadiness({
-        botRequestId,
-        cloudAgentSessionId,
-      })
-    ).resolves.toMatchObject({ status: 'result-error' });
-    await expect(
-      claimBotRequestCloudAgentSessionGroupContinuation({
-        botRequestId,
-        cloudAgentSessionId,
-      })
-    ).resolves.toBe(true);
-  });
-
-  it('orders child session groups by callback step, creation time, and session id', async () => {
-    const botRequestId = await createBotRequest();
-    const spawnGroupId = randomUUID();
-    const firstCloudAgentSessionId = `cas-child-order-first-${randomUUID()}`;
-    const secondCloudAgentSessionId = `cas-child-order-second-${randomUUID()}`;
-    createdCloudAgentSessionIds.add(firstCloudAgentSessionId);
-    createdCloudAgentSessionIds.add(secondCloudAgentSessionId);
-
-    await recordBotRequestCloudAgentSession({
-      botRequestId,
-      spawnGroupId,
-      cloudAgentSessionId: firstCloudAgentSessionId,
-      callbackStep: 3,
-    });
-    await recordBotRequestCloudAgentSession({
-      botRequestId,
-      spawnGroupId,
-      cloudAgentSessionId: secondCloudAgentSessionId,
-      callbackStep: 2,
-    });
-
-    const group = await getBotRequestCloudAgentSessionGroup({
-      botRequestId,
-      cloudAgentSessionId: firstCloudAgentSessionId,
-    });
-
-    expect(group.sessions.map(session => session.cloud_agent_session_id)).toEqual([
-      secondCloudAgentSessionId,
-      firstCloudAgentSessionId,
-    ]);
   });
 
   it('continues linking the legacy Cloud Agent session column', async () => {
